+++ Assets/__Scripts/MapEditor/Grid/LaneInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 91df446b093f5c045b2f57255e25dbd5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
